refine embedding page
